server/requirements.txt
Flask==2.3.3
joblib==1.4.0
scikit-learn==1.4.2
numpy==1.26.4
Flask-Login==0.6.3
Werkzeug==2.3.7
python-dotenv>=1.0.0
gunicorn==20.1.0
flask-jwt-extended==4.6.0
flask_cors==3.0.10
google-genai==1.31.0
pandas